requirements.txt
astroid==2.2.5
Click==7.1.2
ecdsa==0.13.2
Flask==2.0.0
Flask-SQLAlchemy==2.5.0
future==0.17.1
isort==4.3.18
itsdangerous==2.0.1
Jinja2==3.0.0
lazy-object-proxy==1.4.0
MarkupSafe==2.0.1
mccabe==0.6.1
pycryptodome==3.3.1
pylint==2.3.1
python-jose==3.3.0
six==1.12.0
# typed-ast==1.4.2
Werkzeug==2.0.0
wrapt==1.11.1
Flask-Cors==3.0.8
SQLAlchemy==1.4.41
pip~=24.3.1
wheel~=0.41.2
cryptography~=44.0.0
rsa~=4.9
pyasn1~=0.6.1
setuptools~=68.2.0
cffi~=1.17.1
greenlet~=3.1.1
pycparser~=2.22
jupyter~=1.0.0
attrs~=23.2.0
argparse~=1.4.0
pillow~=10.4.0
filelock~=3.15.4
colorama~=0.4.6
blinker~=1.9.0
appdirs~=1.4.4
typing_extensions~=4.12.2
scipy~=1.14.1
psutil~=6.0.0
gevent~=24.11.1
ply~=3.11
decorator~=5.1.1
packaging~=24.1
legacy~=0.1.7
platformdirs~=4.2.2
lxml~=5.3.0
beautifulsoup4~=4.12.3
soupsieve~=2.5
defusedxml~=0.7.1
numpy~=2.0.0
ipython~=8.26.0
pyzmq~=26.0.3
pexpect~=4.9.0
tornado~=6.4.1
comm~=0.2.2
traitlets~=5.14.3
fqdn~=1.5.1
parso~=0.8.4
jedi~=0.19.1
QtPy~=2.4.1
PyQt5~=5.15.10
PyYAML~=6.0.1
sniffio~=1.3.1
arrow~=1.3.0
python-dateutil~=2.9.0.post0
python-dotenv~=1.0.1
pytz~=2024.1
overrides~=7.7.0
numba~=0.60.0
threadpoolctl~=3.5.0
torch~=2.2.2
matplotlib~=3.9.1
mpmath~=1.3.0
sympy~=1.12.1
pooch~=1.8.2
bleach~=6.1.0
webencodings~=0.5.1
tinycss2~=1.3.0
debugpy~=1.8.2
Pygments~=2.18.0
jupyter_client~=8.6.2
prompt_toolkit~=3.0.47
wcwidth~=0.2.13
ptyprocess~=0.7.0
nbformat~=5.10.4
jupyter_core~=5.7.2
nbclient~=0.10.0
fastjsonschema~=2.20.0
jsonschema~=4.22.0
requests~=2.32.3
notebook~=7.2.1
prometheus_client~=0.20.0
Send2Trash~=1.8.3
terminado~=0.18.1
asttokens~=2.4.1
executing~=2.0.1
ipykernel~=6.29.5
jupyter_server~=2.14.1
notebook_shim~=0.2.4
nbconvert~=7.16.4
pandocfilters~=1.5.1
mistune~=3.0.2
jupyterlab_pygments~=0.3.0
qtconsole~=5.5.2
websocket-client~=1.8.0
ipywidgets~=8.1.3
referencing~=0.35.1
idna~=3.7
webcolors~=24.6.0
jsonpointer~=2.1
isoduration~=20.11.0
rich~=13.9.4
jupyter_server_terminals~=0.5.3
anyio~=4.4.0
document~=1.0